% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_school_sheet.R
\name{find_school_sheet}
\alias{find_school_sheet}
\title{Identify the sheet in a FAFSA report that contains school-level data}
\usage{
find_school_sheet(.fn)
}
\arguments{
\item{.fn}{`<chr>` the full path to a FAFSA report spreadsheet}
}
\value{
`<chr>` the name of the sheet
}
\description{
Identify the sheet in a FAFSA report that contains school-level data
}
