% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hercfafsa-package.R
\docType{package}
\name{hercfafsa-package}
\alias{hercfafsa}
\alias{hercfafsa-package}
\title{hercfafsa: Wrangle Federal FAFSA Reports into Tidy Data}
\description{
There are several different kinds of reports available at studentaid.gov. None of them are in tidy format. This package aims to solve that.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/higherX4Racine/hercfafsa}
  \item Report bugs at \url{https://github.com/higherX4Racine/hercfafsa/issues}
}

}
\author{
\strong{Maintainer}: Ben Taft \email{ben.taft@career2cradle.org}

Other contributors:
\itemize{
  \item Higher Expectations for Racine County [copyright holder]
}

}
\keyword{internal}
