% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_status_and_date.R
\name{parse_status_and_date}
\alias{parse_status_and_date}
\title{Split a column containing both FAFSA status and date information}
\usage{
parse_status_and_date(.x, .chimera_column)
}
\arguments{
\item{.x}{`<obj>` the data frame with a chimeric column}

\item{.chimera_column}{`<chr>` the name of the chimeric column}
}
\value{
a tibble like `.x`, without `.chimera_column`,with two extra columns
\describe{
  \item{Status}{`<chr>` Either "Submitted" or "Complete"}
  \item{Date}{`<date>` When the counts end}
}
}
\description{
Split a column containing both FAFSA status and date information
}
