% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_school_data.R
\name{read_raw_school_data}
\alias{read_raw_school_data}
\title{Pull in school-level data from a FAFSA report}
\usage{
read_raw_school_data(.fn, .sheet)
}
\arguments{
\item{.fn}{`<chr>` the full path to the report spreadsheet}

\item{.sheet}{`<chr>` the sheet in that file with school-level data}
}
\value{
`<obj>` a data frame with a variable number of columns
}
\description{
Pull in school-level data from a FAFSA report
}
