% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_school_data.R
\name{read_school_data}
\alias{read_school_data}
\title{Import data about FAFSA applications at specific schools}
\usage{
read_school_data(.fn)
}
\arguments{
\item{.fn}{`<chr>` the full path to the FAFSA report spreadsheet}
}
\value{
`<obj>` a tibble with seven or eight columns
\describe{
  \item{School Code}{`<chr>` the numeric code of the school, only in later years}
  \item{Name}{`<chr>` the name of the school}
  \item{City}{`<chr>` the municipality where the school is located}
  \item{State}{`<chr>` the state where the school is located}
  \item{Cycle}{`<chr>` which FAFSA cycle the data are part of}
  \item{Status}{`<chr>` either "Submitted" or "Complete"}
  \item{Date}{`<date>` the end of the month the data correspond to}
  \item{Applications}{`<int>` the number of applications}
}
}
\description{
Import data about FAFSA applications at specific schools
}
