% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{STUDENT_STATUS_REPORT_FORMATS}
\alias{STUDENT_STATUS_REPORT_FORMATS}
\title{reports about individual student status come in several formats.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 5 rows and 4 columns.
}
\usage{
STUDENT_STATUS_REPORT_FORMATS
}
\description{
reports about individual student status come in several formats.
}
\keyword{datasets}
