% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_school_headers.R
\name{wrangle_school_headers}
\alias{wrangle_school_headers}
\title{Convert raw header cells into a tidy table}
\usage{
wrangle_school_headers(.raw_headers)
}
\arguments{
\item{.raw_headers}{results from running [read_raw_school_headers()]}
}
\value{
`<obj>` a tibble with three columns
\describe{
  \item{Cycle}{`<chr>` The FAFSA cycle that this column reports on}
  \item{Date}{`<date>` The date that this column corresponds to}
  \item{Status}{`<chr>` Either "Submitted" or "Complete"}
}
}
\description{
Convert raw header cells into a tidy table
}
